Updated / pushed details to the masterbranch.
meta-nvidia
===========
This layer adds support for NVidia graphics cards using the upstream
binary drivers. It's a pile of dirty hacks and will likely break your
build in new and exciting ways. This layer makes no attempt to support
any platform other than x86_64. It has only been tested with the pyro
branch of yocto.

To use this layer, include it in your bblayers.conf and add the
following to your BSP, distro, or local config:

  PREFERRED_VERSION_mesa = "17.%"
  PREFERRED_PROVIDER_virtual/libgl = "libglvnd"
  PREFERRED_PROVIDER_virtual/libgles1 = "libglvnd"
  PREFERRED_PROVIDER_virtual/libgles2 = "libglvnd"
  PREFERRED_PROVIDER_virtual/egl = "libglvnd"
  PREFERRED_PROVIDER_virtual/mesa = "libglvnd"
  KERNEL_MODULE_AUTOLOAD_append = " nvidia nvidia-drm nvidia-modeset nvidia-uvm"
  XSERVER = " \
      ${XSERVER_X86_BASE} \
      ${XSERVER_X86_EXT} \
      ${XSERVER_X86_MODESETTING} \
      nvidia"
  DISTRO_FEATURES_append = " x11 opengl"
  DISTRO_FEATURES_remove = " wayland"

Optionally, add the nvidia-docker and nvidia-docker-compose packages if you
desire that functionality.


nvidia
======
The nvidia source package builds the kernel modules and installs the
userspace libraries from NVidia's binary package. The userspace
libraries depend on libglvnd, which attempts to wrap all GL calls and
redirect them to the correct libraries at runtime. This build mostly
supersedes mesa, except for it's GLX implementation.


nvidia-docker
=============
The nvidia-docker package is included to ease the pain of running
hardware accelerated applications inside docker containers. Building
this package depends on your host system having a working docker
service. meta-virtualization is only a dependency if nvidia-docker
support is needed.


Dependencies
============

This layer depends on:

  URI: git://git.yoctoproject.org/poky
  layers: meta
  branch: pyro

  URI: git://git.yoctoproject.org/meta-openembedded
  layers: meta-oe
  branch: pyro

  URI: git://git.yoctoproject.org/meta-virtualization
  layers: meta-virtualization
  branch: pyro